feat: updated add fluent network button
